+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Adding `or` validator
+
+### Fixed
+
+- Fixing translation in global validators
+
+
 ## [1.2.0] - 2019-01-25
 
 ### Added